Add memoized fibonacci.

+++ src/main/kotlin/com/github/phphavok/kotlinalgorithms/algorithms/fibonacci/IFibonacci.kt
@@ -1,5 +1,7 @@
 package com.github.phphavok.kotlinalgorithms.algorithms.fibonacci
 
+import java.math.BigInteger
+
 interface IFibonacci {
-    fun fibonacci(n: Long): Long
+    fun fibonacci(n: UInt): BigInteger
 }
